Refactored PlanarTileMap. Using std::pair to hold data lead to really confusing code. I'm now using a struct instead.
Fixed algorithm to generate texture names for tilings. I was selecting the wrong view before.

Algorithm to generate texture names does not generate textures which are not visible from the center fundamental copy.

Removed a bunch of print statements which were not useful.

There is still a bug where one triangle is not renderered. I'm about to start looking into it.